+++ dwds/test/expression_compiler_service_test.dart
@@ -7,6 +7,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:dwds/dwds.dart';
+import 'package:dwds/src/services/expression_compiler.dart';
 import 'package:dwds/src/utilities/shared.dart';
 import 'package:http_multi_server/http_multi_server.dart';
 import 'package:path/path.dart' as p;
@@ -139,7 +140,8 @@ void main() async {
       expect(output.stream,
           emitsThrough(contains('[INFO] ExpressionCompilerService: Stopped.')));
 
-      var result = await service.updateDependencies({'try': 'try.full.dill'});
+      var result = await service
+          .updateDependencies({'try': ModuleInfo('try.full.dill', 'try.dill')});
       expect(result, true, reason: 'failed to update dependencies');
 
       final compilationResult = await service.compileExpressionToJs(